Parse some basic LaTeX input
Summary:
For the Desmos/MathQuill integration we will need to handle
their output, which is in LaTeX.

As it is, this does not handle all the possible things that can come out
of MathQuill, since it allows expressions like `\sum _{i=n}^5i`, so we
will either have to handle everything that could reasonably come up, or
restrict the things that can be put in.

I wasn't completely sure which token \frac should be a part of
(parser-generator.js L167), primitive seemed like an ok choice.

Also, I think the quadruple-backslashes in parser-generator are a bit
obnoxious, but I think they are needed just based on how the parser is
generated.

Test Plan:
* Run a local server
* Open test.html
* Verify all tests pass

Reviewers: alex

Reviewed By: alex

Differential Revision: http://phabricator.khanacademy.org/D6002