First pass at removing TeX output
Summary:
Since Desmos/MathQuill will be handling the formatting of any input, it
should be safe to remove the TeX output.

This commit removes the TeX rendering, the associated tests, and some of
the hints.

I believe the remaining hints are necessary for both parsing properly and for
determining if an expression is simplified.

I removed the 3 different constants for the different kinds of implicit
negation, and replaced them with just one (Num.ImplicitNeg), which has the hint
"implicit". I also added an explanation for each of the remaining hints and
what they do.

Test Plan:
  * Open test.html
  * Verify all tests pass

Reviewers: alex

Reviewed By: alex

Differential Revision: http://phabricator.khanacademy.org/D6072