Create React version of the experimenter
Summary:
I redid the experimenter in React so it would be a bit easier to manage
and also to get a better handle on React in general.

I did it on the tex_input branch rather than the one that has the first
version of the experimenter since in order for the MathQuill stuff to
work we need to be able to parse their TeX, so once we merge it back in
we can remove the old one.

Test Plan:
  * Open up experimenter/index.html and play around with it

Reviewers: alex, jack

Reviewed By: alex

Differential Revision: http://phabricator.khanacademy.org/D6096