Change method of calculating difference
Summary:
Previously, `3^-x` and `(1/3)^x` were considered equivalent. This was
caused by the difference between the numbers being quite large for
certain values of x. This commit changes the method of calculating the
delta so it is calculated relatively instead of absolutely, and also
adds some negative test cases to make sure this isn't too lenient as
well.

Test Plan:
* Open test.html
* See all tests pass!

Reviewers: alex

Reviewed By: alex

Differential Revision: http://phabricator.khanacademy.org/D6324