Remove support for parsing '/=' and '!=' as "not equals"
'/=' was always a weird way to do it, and too close to the much clearer '=/='
'!=' is ambiguous: does '3!=6' mean "3 is not 6" or "3 factorial is 6"?
(We don't actually parse factorials at the moment, but we likely will soon.)

Test Plan:
Confirmed that none of our existing equations use either of these, then
removed the relevant test cases.

Auditors: jack