Work on fixing the division form problem.
... where 2/(3*4) and 2/(4*3) don't count as the same form.

The change in handleDivide is the core of my idea for fixing it.
Basically, we float all the negatives to the front multiplied fractions.

Unfortunately it didn't work quite that well in practice. Many things
were printed differently, which makes sense, but then `Mul.completeParse`
breaks because there's now a case in which the only term is a
`Rational`, so the old `completeParse` returned a `Mul` with `terms`
just set to that `Rational`. We may want to pull in this change anyway.

`Mul.tex` also had what I believe is a legit bug, though maybe not -
it's hard to say if I broke any implicit invariants. Check out this
test:

> -1/1/x renders as \frac{-1}{x}
> Expected: "\frac{-1}{x}"
> Result: "\frac{1}{x}"

... which I was trying to fix with the negatives change.

Anyway, I think the proper fix is to entirely rethink `handleDivide`,
but that's a larger discussion. We may be able to make use of the
smaller fixes anyway.

https://app.asana.com/0/20879068182935/22467996873952

Auditors: alex