Add support for nth roots.
Summary:
Support ticket: https://app.asana.com/0/27216215224639/38156337055613

We now support the syntax `sqrt[3]{4}` to mean the cube root of 4. The
degree provided can be any expression, but when outputting to tex, if
the degree is not an integer the output will be a little funky. Here's
some expressions + their tex output from KAS:

    > sqrt[3]{4}
    \sqrt[3]{4}
    > sqrt[x]{4}
    4^{\frac{1}{x}}

This issue will affect users of the non MathQuill expression widget in
Perseus, but aside from some initial confusion they shouldn't have a
terribly hard time.

Notably, I changed the behavior of the options.dynamic property of the
`asTex` method to not swap out square brackets with `\left[` and
`\right[`. I played around with the old expression widget and couldn't
figure out anywhere this broke things, so I'm hoping it doesn't hurt
anything. This was done because `\sqrt\left[\right]{x}` is not valid
syntax to MathJax.

Test Plan: Run tests in /test.html.

Reviewers: alex

Reviewed By: alex

Differential Revision: https://phabricator.khanacademy.org/D18923