Add special case for Math.pow with negative bases and fractional exponents
Summary:
It turns out that Math.pow returns NaN whenever it's provided with
both a negative base and a fractional exponent (since roots are close together
and floating point numbers are only so precise). This is an issue when users
deal with, for example, cube roots, and cbrt(-5) is evaluated, in the end, as:

    Math.pow(-5, 0.33333333) // NaN

which makes cbrt(-5) unequal to -cbrt(5) and other acceptable results. This
diff uses information that we have about the base and exponent to determine
a subset of these cases.

Test Plan:
- Added multiple new test cases to `test.html` (run with `grunt` and then open `test.html`).
- Verified that `\sqrt[3]{-5}` parses correctly in `experimenter.html`.

Reviewers: alex

Reviewed By: alex

Differential Revision: https://phabricator.khanacademy.org/D21792