Add a method for collecting all constants in an expression
Summary:
For our new math-input, we need to be able to collect all of the constants and variables in an
expression, so that we can expose them as keys on the keypad. This diff adds a `getConsts`
method that mirrors the existing `getVars` method.

Test Plan: Run the new unit tests.

Reviewers: alex

Reviewed By: alex

Subscribers: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27242