Updated expression editor to recognize exponents and radicals as structurally differenteven if evaluated the same.
Summary:
The problem was that the editor accepted x^(1/8) as the same as rt[8](x) even when same form was checked. This meant that users could cheat on exercises that tested the difference between radicals and exponents. I fixed it by updating the Pow variant with a field that notes whether it was created from a radical or exponent (which are parsed differently).
Asana task: https://app.asana.com/0/253524899192327/228149900403726

Test Plan:
I verified at this link: math/algebra/rational-exponents-and-radicals/rational-exponents-intro/e/understanding-fractional-exponents
I tried 4 main test cases. If the prompt was x^(1/8), I tried:
 - exponent with correct variable but wrong number: x^(1/5)
 - exactly the same as prompt: x^(1/8)
 - radical with wrong number: rt[5](x)
 - radical with right number: rt[8](x)
Only the last test case should be correct.

Reviewers: emily

Reviewed By: emily

Subscribers: alex

Differential Revision: https://phabricator.khanacademy.org/D36445