Update to grunt 1 and qunit 2
There are security vulnerabilities in the earlier versions of these
modules. The QUnit change required massive changes to the test file
(asserts are no longer global). I was able to get most everything
working, but there are a few tests that still fail. There are failing
tests in master though, so...