Change expression widget's sameForm method to use .tex()
In the Khan Academy web app, we have 2 exercises that require the answer given by the user to be in a specific form. One of those exercises is working. The other is not.

I'm not totally clear on the specifics of how KaTex works, but for the exercise that is not working, the parsed form looks the same no matter what format the answer is in. That means that no matter what form the answer is in, it's always passing the form comparison. The `sameForm` method is currently taking the parsed expression and calling `.strip()`. Changing it to use `.tex()` makes the form comparison work for both exercises.

This change should not affect exercises that don't care what form the answer is in. Only these two exercises are known to use that option.

Test plan:
I made this change in the webapp repo in third_party/javascript-khansrc/kas,
navigate to both `/math/algebra2/x2ec2f6f830c9fb89:exp/x2ec2f6f830c9fb89:rational-exp/e/understanding-fractional-exponents` and `/math/algebra2/x2ec2f6f830c9fb89:exp/x2ec2f6f830c9fb89:rational-exp/e/exponents_3`,
did a bunch of questions where I first copied the question format then used the format they were looking for,
and verified that the question format was considered incorrect and the expected format was considered correct.